add missing static_cast

+++ src/sf-as-network.cpp
@@ -98,7 +98,7 @@ Rcpp::List rcpp_sf_as_network (const Rcpp::List &sf_lines,
         Rcpp::NumericMatrix gi = (*g);
         size_t rows = static_cast <size_t> (gi.nrow () - 1);
         nrows += rows;
-        if (ngeoms < ow.size ())
+        if (ngeoms < static_cast <unsigned int> (ow.size ()))
         {
             //if (!(ow [ngeoms] == "yes" || ow [ngeoms] == "-1"))
             if (ow [ngeoms] == "yes" || ow [ngeoms] == "true")
